put sub-components back into Dropdown.js files & create 2 files for SideNav and TopNav
