+++ src/components/Header/outlineStyles.js
@@ -6,6 +6,14 @@ const outlineStyles = css`
     outline: ${remcalc(4)} solid ${props => props.theme.colors.vibrant};
   }
 
+  ${props =>
+    props.themeVariation === 'dark' &&
+    css`
+      &:focus {
+        outline: ${remcalc(4)} solid ${props => props.theme.colors.white};
+      }
+    `}
+
   &:active {
     outline: none;
   }
